All the code are organized with proper structure
