+++ frontend/src/app/App.jsx
@@ -1,0 +1,13 @@
+import './App.css'
+
+function App() {
+
+  return (
+    <>
+      
+      <h1>Vite + React</h1>
+    </>
+  )
+}
+
+export default App
