+++ frontend/src/app/layout/Main.jsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+
+const Main = () => {
+    return (
+        <div>
+            <Outlet></Outlet>
+        </div>
+    );
+};
+
+export default Main;
