--- frontend/src/components/Root/Root.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-
-const Root = () => {
-    return (
-        <div>
-            <Outlet></Outlet>
-        </div>
-    );
-};
-
-export default Root;
