--- frontend/src/components/SignUp/SignUp.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import RegisterForm from '../RegisterForm/RegistrationForm';
-import Footer from '../Footer/Footer';
-
-const SignUp = () => {
-    return (
-        <div>
-            <RegisterForm></RegisterForm>
-            <Footer></Footer>
-        </div>
-    );
-};
-
-export default SignUp;
