+++ frontend/src/features/landingPage/pages/home/Home.jsx
@@ -1,0 +1,15 @@
+import Navbar from '../../../../shared/components/navbar/Navbar';
+import HomeLand from '../../components/home/HomeLand';
+import Footer from '../../../../shared/components/footer/Footer';
+
+const Home = () => {
+    return (
+        <div>
+            <Navbar></Navbar>
+            <HomeLand></HomeLand>
+            <Footer></Footer>
+        </div>
+    );
+};
+
+export default Home;
