--- frontend/src/main.jsx
@@ -1,38 +1,0 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import {
-  createBrowserRouter,
-  RouterProvider,
-} from "react-router-dom";
-import Root from './components/Root/Root';
-import Home from './components/Home/Home';
-import MachineDetailsPage from './components/MachineDetailsPage/MachineDetailsPage';
-import SignUp from './components/SignUp/SignUp';
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root></Root>,
-    children: [
-      {
-        path: "/",
-        element: <Home></Home>,
-      },
-      {
-        path: "machine_details/",
-        element: <MachineDetailsPage></MachineDetailsPage>,
-      },
-      {
-        path: "signup/",
-        element: <SignUp></SignUp>,
-      },
-    ],
-  },
-]);
-
-createRoot(document.getElementById('root')).render(
-  <StrictMode>
-    <RouterProvider router={router} />
-  </StrictMode>,
-)
