+++ frontend/src/shared/components/footer/Footer.jsx
