+++ frontend/tailwind.config.js
@@ -1,30 +1,14 @@
-// /** @type {import('tailwindcss').Config} */
-// export default {
-//   content: [
-//     "./index.html",
-//     "./src/**/*.{js,ts,jsx,tsx}",
-//   ],
-//   theme: {
-//     extend: {},
-//   },
-//   plugins: [
-//     require('daisyui'),
-//   ],
-// }
-
-
 /** @type {import('tailwindcss').Config} */
-import daisyui from 'daisyui';
-
 export default {
   content: [
-    './index.html',
-    './src/**/*.{js,ts,jsx,tsx}',
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
   },
-  plugins: [daisyui],
-};
-
+  plugins: [
+    require('daisyui'),
+  ],
+}
